+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 2.7.0 (unreleased)
 ------------------
 
+- #3005 Remove unused getWorksheetTemplateUID/URL worksheet catalog columns
+- #3004 Speed up the creation of worksheets with a worksheet template
 - #3008 Filter sidebar root folders by permission instead of by catalog
 - #3007 Fix empty sticker template selection when printing stickers from a listing
 - #3003 Fix colliding reference analyses group IDs within the same transaction